% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{logL_gammamix}
\alias{logL_gammamix}
\title{logL_gammamix}
\usage{
logL_gammamix(x, w, shape, rate)
}
\arguments{
\item{x}{numeric vector of quantile with at least length 2}

\item{w}{numeric vector of mixture weights}

\item{shape}{numeric vector of shape parameters}

\item{rate}{numeric vector of rate parameters}
}
\description{
log-likelihood of a mixture of gamma distributions
}
\keyword{internal}
