% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_functions.R
\name{n_ttest}
\alias{n_ttest}
\title{Sample size of a t-test}
\usage{
n_ttest(variance, alloc, delta, sig_level, power)
}
\arguments{
\item{variance}{assumed within-group variance}

\item{alloc}{numeric vector of sample size proportion within group}

\item{delta}{assumed treatment effect}

\item{sig_level}{level of significance}

\item{power}{target power}
}
\value{
sample size of the one-sided two-sample t-test
}
\description{
Sample size calculation of a one-sided two-sample t-test from
the non-central t-distribution. The variances are equal between the groups,
the sample sizes must not be equal.
}
\examples{
n_ttest(variance = 1, alloc = c(0.5, 0.5), delta = 0.5,
        sig_level = 0.025, power = 0.8)
}
