% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmixinvgamma.R
\name{qmixinvgamma}
\alias{qmixinvgamma}
\title{qmixinvgamma}
\usage{
qmixinvgamma(p, w, shape, rate)
}
\arguments{
\item{p}{vector of probabilities}

\item{w}{weights which sum up to one}

\item{shape}{shape parameter}

\item{rate}{rate parameter}
}
\description{
Quantiles of a mixture of invserse gamma distributions
}
\examples{
n <- c(800000, 400000)
x <- c(rgamma(n[1], shape = 3, rate = 2), rgamma(n[2], shape = 5, rate = 4))
quantile(1/x, probs = 0.75)
qmixinvgamma(p = 0.75, w = n/sum(n), shape = c(3, 5), rate = c(2,4))
}
