% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_var_prior.R
\name{update_var_prior}
\alias{update_var_prior}
\title{Update the MAP prior of the variance}
\usage{
update_var_prior(prior, sample_var, degf)
}
\arguments{
\item{prior}{a list with elements w, shape, rate specifying the parameters
of the mixture of inverse Gamma distributions.}

\item{sample_var}{a number specifying the observed sample variance}

\item{degf}{a number specifying the degrees of freedom from the sample
variance}
}
\description{
This function updates MAP prior of the variance, a mixture
of inverse Gamma distributions, based on an estimated of the sample variance.
}
