% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varianceHAMD}
\alias{varianceHAMD}
\title{Sample variances of Hamilton Rating Scale for Depression (HAMD) score
from eleven studies}
\format{An object of class \code{data.frame} with 11 rows and 3 columns.}
\usage{
varianceHAMD
}
\description{
A data set containing the pooled sample variances and the
respective degrees of freedom for the endpoint HAMD score
from eleven studies as reported in Analysis 2.3 by Linde et al.
}
\references{
Linde K, Berner MM, Kriston L.
John’s wort for major depression.
The Cochrane Library 2008; 4:CD000448
}
\keyword{datasets}
