% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varianceSPB}
\alias{varianceSPB}
\title{Sample variances of systolic blood pressure from twelve studies}
\format{An object of class \code{data.frame} with 12 rows and 3 columns.}
\usage{
varianceSPB
}
\description{
A data set containing the pooled sample variances and the
respective degrees of freedom for the endpoint systolic blood pressure
from twelve studies as reported in Analysis 1.1 by Glynn et al.
}
\references{
Glynn LG, Murphy AW, Smith SM, Schroeder K, Fahey T.
Interventions used to improve control of blood pressure in patients with hypertension.
The Cochrane Library 2010; 3:CD005182.
}
\keyword{datasets}
