% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmeta.R
\name{varmeta}
\alias{varmeta}
\title{MAP approach for variances}
\usage{
varmeta(sample_var, degf, mu_mean, n.iter = 10000, ...)
}
\arguments{
\item{sample_var}{numeric vector with the sample variances from historical
clinical trials}

\item{degf}{numeric vector with the degrees of freedom of the sample variances}

\item{mu_mean}{mean of the normal prior for the mean}

\item{n.iter}{number of iterations to monitor in the R function \code{coda.samples}}

\item{...}{Additional arguments for the R functions \code{coda.samples} and
\code{jags.model}}
}
\value{
An \code{mcmc.list} object with the posterior predictive for the variance
(\code{map_variance}), the posterior for the mean of the log variance (\code{mu}),
and the between-trial standard deviation of the log-variance (\code{tau})
}
\description{
Performs a random effects meta-analysis for variances.
}
\details{
The function \code{varmeta} performs a random effect meta-analysis
for variance.
The model is as proposed by Schmidli et al., i.e. the log-variance are
modeled using a normal hierarchical model with normal prior of the mean
and a half-normal prior for the (between-trial) standard deviation.
}
\references{
Schmidli H, Neuenschwander B, Friede T.
Meta-analytic-predictive use of historical variance data for the design and
analysis of clinical trials.
Computational Statistics & Data Analysis (2016).
}
